Fix Windows build. Small tweaks.
Confirmed that DIFFERENCE_SYNC really does help stabalize the waveform, and also that a certain linear interpolation on a buffer related to DIFFERENCE_SYNC also improves stabilization. With FFT_SYNC and DIFFERENCE_SYNC, the visualizer is able to get a firm grip on many waveforms and hold them steady in the viewport :). It works good on windows currently. Linux WNIP (work not in progress).