Fix error in usage of framebuffers.
I noticed that the multirender was not working when rendering on my gtx 1060 gfx card. I confirmed that the one of the textures associated with each buffer was not being rendered to. To show that a texture was not being rendered to I used the image buffer's fragment shader to draw the A buffer's contents only on odd frames. If the A buffer was rendering to both textures, then the image fragment shader would always have non black output. However, the image's fragment shader would only have non black output for odd frames. This shows that something was wonky with which textures were being rendered to for a buffer.

I installed some debug checks and noticed that the driver did not like the way I was using textures. I could not find the exact cause of the bug but reworking how I deal with textures and texture units has solved the problem. I think the issue may have been linked with glFramebufferTexture2D and GL_COLOR_ATTACHMENTs.